+++ Aurora/Aurora/net/ethernet.cpp
@@ -46,6 +46,6 @@ void ethernet_send (void* data, size_t len, uint16_t type, uint8_t* dest) {
 	uint8_t *src_mac = AuGetNetTuple()->mac;
 	memcpy(packet->src,src_mac,6);
 	packet->type_len = htons(type);
-	AuNetWrite((uint8_t*)packet,total_sz);
+	AuNetWrite((uint64_t*)packet,total_sz);
 	free(packet);
 }
