+++ Aurora/Aurora/utils/pipe.cpp
@@ -34,7 +34,7 @@ void pipe_read (vfs_node_t *file, uint64_t* buffer,uint32_t length) {
 }
 
 
-void pipe_write (vfs_node_t *file, uint8_t* buffer, uint32_t length) {
+void pipe_write (vfs_node_t *file, uint64_t* buffer, uint32_t length) {
 	pipe_t *p = (pipe_t*)file->device;
 	for (int i = 0; i < length; i++) {
 		circular_buf_put(p->buf, buffer[i]);
