+++ Aurora/include/fs/vfs.h
@@ -57,7 +57,7 @@ struct _vfs_node_;
 typedef _vfs_node_ (*open_callback) (_vfs_node_ *node, char* path);
 typedef void (*read_callback) (_vfs_node_ *file, uint64_t* buffer,uint32_t length);
 typedef void (*read_block_callback) (_vfs_node_ *file, uint64_t* buffer);
-typedef void (*write_callback) (_vfs_node_ *file, uint8_t* buffer, uint32_t length);
+typedef void (*write_callback) (_vfs_node_ *file, uint64_t* buffer, uint32_t length);
 typedef int (*ioquery_callback) (_vfs_node_ *file, int code, void *arg);
 
 //! The File Node -- everything is file
@@ -108,7 +108,7 @@ AU_EXTERN AU_EXPORT void vfs_mount (char *path, vfs_node_t *node, vfs_entry* ent
 extern void vfs_lsdir (char* path);
 AU_EXTERN AU_EXPORT vfs_node_t openfs (vfs_node_t *node, char* path);
 AU_EXTERN AU_EXPORT void readfs (vfs_node_t *node, vfs_node_t *file, uint64_t* buffer, uint32_t length);
-AU_EXTERN AU_EXPORT void writefs (vfs_node_t *node, vfs_node_t *file, uint8_t *buffer, uint32_t length);
+AU_EXTERN AU_EXPORT void writefs (vfs_node_t *node, vfs_node_t *file, uint64_t *buffer, uint32_t length);
 AU_EXTERN AU_EXPORT void readfs_block (vfs_node_t* node, vfs_node_t *file,  uint64_t *buffer);
 AU_EXTERN AU_EXPORT int  vfs_ioquery (vfs_node_t *node, int code, void* arg);
 #endif
